Add: code cleaning (comments and move jwt_decode to jwt_utils)
